feat: D9 part1

+++ 2024/src/main/kotlin/Day.kt
@@ -1,6 +1,8 @@
+import java.math.BigInteger
+
 interface Day {
-    fun solvePart1(debug : Boolean = false): Int
-    fun solvePart2(debug : Boolean = false): Int
+    fun solvePart1(debug : Boolean = false): Long
+    fun solvePart2(debug : Boolean = false): Long
 
     fun getDay(): Int {
         return this::class.java.simpleName.substring(3).toInt()
