+++ 2024/src/main/kotlin/days/Day3.kt
@@ -7,20 +7,20 @@ class Day3 : Day {
     private val mulRegex = Regex("mul\\((?<leftOperand>\\d{1,3}),(?<rightOperand>\\d{1,3})\\)")
     private val doToDontRegex = """mul\((?<leftOperand>\d{1,3}),(?<rightOperand>\d{1,3})\)|do(n't)?\(\)""".toRegex()
 
-    override fun solvePart1(debug: Boolean): Int {
+    override fun solvePart1(debug: Boolean): Long {
        return  mulRegex.findAll(input).map { it.groups }.map {
-           (it["leftOperand"]!!.value.toInt() ?: 1) * it["rightOperand"]!!.value.toInt()
+           (it["leftOperand"]!!.value.toLong() ?: 1) * it["rightOperand"]!!.value.toLong()
        }.sumOf { it }
     }
-    override fun solvePart2(debug: Boolean): Int {
+    override fun solvePart2(debug: Boolean): Long {
        return doToDontRegex.findAll(input).map{it.groups}.fold(
-            Pair(true,0), {acc, group ->
+            Pair(true,0.toLong()), {acc, group ->
                 when (group[0]?.value){
                     "do()" ->  Pair(true, acc.second)
                     "don't()" -> Pair(false, acc.second)
                     else -> {
                         if (acc.first) {
-                            Pair(acc.first, acc.second + group["leftOperand"]!!.value.toInt() * group["rightOperand"]!!.value.toInt())
+                            Pair(acc.first, acc.second + group["leftOperand"]!!.value.toLong() * group["rightOperand"]!!.value.toLong())
                         } else {
                             acc
                         }
