+++ 2024/src/main/kotlin/days/Day7.kt
@@ -5,7 +5,7 @@ import java.math.BigInteger
 import kotlin.math.pow
 
 
-data class Equation(val result: BigInteger, val operands: List<BigInteger>)
+data class Equation(val result: Long, val operands: List<Long>)
 class Day7 : Day {
 
     private val equations: MutableList<Equation> = mutableListOf()
@@ -14,26 +14,26 @@ class Day7 : Day {
         val input = getInput()
         input.split(System.lineSeparator()).forEach { line ->
             val equation = line.trim().split(":")
-            val result = equation[0].trim().toBigInteger()
-            val operands = equation[1].trim().split(" ").map { it.trim().toBigInteger() }
+            val result = equation[0].trim().toLong()
+            val operands = equation[1].trim().split(" ").map { it.trim().toLong() }
             equations.add(Equation(result, operands))
         }
     }
 
     private fun solveEquation(
         equation: Equation,
-        operators: List<Pair<String, (BigInteger, BigInteger) -> BigInteger>>,
+        operators: List<Pair<String, (Long, Long) -> Long>>,
         debug: Boolean = false
-    ): BigInteger {
+    ): Long {
         val numberOfPossibleOperator = equation.operands.size - 1
         for (i in 0 until operators.size.toDouble().pow(numberOfPossibleOperator.toDouble()).toInt()) {
-            val operatorCombination = mutableListOf<Pair<String, (BigInteger, BigInteger) -> BigInteger>>()
+            val operatorCombination = mutableListOf<Pair<String, (Long, Long) -> Long>>()
             for (j in 0 until numberOfPossibleOperator) {
                 val operatorIndex = (i / operators.size.toDouble().pow(j.toDouble()).toInt()) % operators.size
                 operatorCombination.add(operators[operatorIndex])
             }
             var successfulEquation = ""
-            val result = equation.operands.foldIndexed(0.toBigInteger()) { index, acc, operand ->
+            val result = equation.operands.foldIndexed(0.toLong()) { index, acc, operand ->
                 if (index == 0) {
                     if (debug)
                         successfulEquation = "$operand"
@@ -50,24 +50,24 @@ class Day7 : Day {
                 return result
             }
         }
-        return 0.toBigInteger()
+        return 0
     }
 
-    override fun solvePart1(debug: Boolean): Int {
+    override fun solvePart1(debug: Boolean): Long {
         val operators = listOf(
-            "+" to { a: BigInteger, b: BigInteger -> a + b },
-            "*" to { a: BigInteger, b: BigInteger -> a * b }
+            "+" to { a: Long, b: Long -> a + b },
+            "*" to { a: Long, b: Long -> a * b }
         )
         println("Result part 1 : ${equations.sumOf { solveEquation(it, operators) }}")
         return 0;
     }
 
 
-    override fun solvePart2(debug: Boolean): Int {
+    override fun solvePart2(debug: Boolean): Long {
         val operators = listOf(
-            "+" to { a: BigInteger, b: BigInteger -> a + b },
-            "*" to { a: BigInteger, b: BigInteger -> a * b },
-            "||" to { a: BigInteger, b: BigInteger -> ("${a}${b}".toBigInteger()) }
+            "+" to { a: Long, b: Long -> a + b },
+            "*" to { a: Long, b: Long -> a * b },
+            "||" to { a: Long, b: Long -> ("${a}${b}".toLong()) }
         )
         println("Result part 2 : ${equations.sumOf { solveEquation(it, operators) }}")
         return 0;
